wanshan

+++ PHP/Service/GiveService/UP.php
@@ -10,14 +10,20 @@ require_once '../../DAO/DAO.php';
 require_once '../UserService/Buffer.php';
 require_once '../UserService/Upload.php';
 if (isset($_SESSION['userid'])) {
-    $src = $_FILES['file']['tmp_name'];
-    $mm = get_md5_string();
-    $N = $mm . '.png';
-    $path = 'http://7xrqhs.com1.z0.glb.clouddn.com/' . $mm . '.png';
-    upload($src, $N, 'give');
-    unlink($src);
-    $_SESSION['bjbj'] = $path;
-    echo "1";
+    $userid = $_SESSION['userid'];
+    if (empty($_SESSION['jwxtnumber'])) {
+        echo '请完善个人信息';
+        die();
+    } else {
+        $src = $_FILES['file']['tmp_name'];
+        $mm = get_md5_string();
+        $N = $mm . '.png';
+        $path = 'http://7xrqhs.com1.z0.glb.clouddn.com/' . $mm . '.png';
+        upload($src, $N, 'give');
+        unlink($src);
+        $_SESSION['bjbj'] = $path;
+        echo "1";
+    }
 }
 function get_md5_string()
 {
